Fix to int

+++ src/Modules/Dlink/InterfacesTrait.php
@@ -170,7 +170,7 @@ trait InterfacesTrait
                 strpos($resp->getValue(), "ch") !== 0
             ) {
                 $indexes[Helper::getIndexByOid($resp->getOid())] = [
-                    'id' =>  Helper::getIndexByOid($resp->getOid()),
+                    'id' =>  (int)Helper::getIndexByOid($resp->getOid()),
                     'name' => $resp->getValue(),
                     '_key' => Helper::getIndexByOid($resp->getOid()),
                     'type' => $types[Helper::getIndexByOid($resp->getOid())],
